refactor: move launcher model gating to runtime
